separate component logic and tests
