+++ core/template/files.go
@@ -1,0 +1,1 @@
+package template
