+++ core/template/files_test.go
@@ -1,0 +1,1 @@
+package template
